# Speed up permissions cache lookups