Uses `$collection->first(...)` instead of `$collection->where(...)->where(...)->first()` to avoid creating two temporary arrays and stop as soon as a match is found.

In my project - PHP 7.4, Xdebug and Debugbar enabled, and 380 permissions - it reduced the `Permission::findByName()` time from 48.54ms average to 6.19ms. On a page with 48 permissions checks, that reduced the total find time from 2330ms to 297ms, and the overall page load **from 3.4 seconds to 1.1 seconds**.