@davejamesmiller 2020-09-25
> If you feel this pull request deserves some attention from my human colleagues feel free to reopen it.

I don't have permission to reopen it? But yes I think it's still an issue.