@drbyte 2020-11-26
@davejamesmiller 
Thanks for this.

After merging with recent updates and exploring your proposed changes, and reflecting on prior discussions around the `clone` and other things, I wonder if the same results can be accomplished, without negative side effect, with reducing this whole PR to just the following combined change?

```diff
   public function getPermissions(array $params = []): Collection
    {
        if ($this->permissions === null) {
            $this->permissions = $this->cache->remember(self::$cacheKey, self::$cacheExpirationTime, function () {
                return $this->getPermissionClass()
                    ->with('roles')
                    ->get();
            });
        }
        
-        $permissions = clone $this->permissions;
-        
-        foreach ($params as $attr => $value) {
-            $permissions = $permissions->where($attr, $value);
-        }
-        
-        return $permissions;

+        return $this->permissions->filter(static function ($permission) use ($params) {
+            foreach ($params as $attr => $value) {
+                if ($permission->getAttribute($attr) != $value) {
+                    return false;
+                }
+            }
+
+            return true;
+        });
    }
```

Curious your thoughts, and whether it still improves your benchmark results similarly to your original results.