@davejamesmiller 2020-11-27
Hi @drbyte 

Thanks for taking a look.

Updated benchmarks for a full page load on the same app as above are:

- `master` - 3.3 seconds
- `speed-up-cache` (my PR) - 1.0 seconds
- [`speed-up-cache-2`](https://github.com/davejamesmiller/laravel-permission/commit/102a37d90bbf1520af3d05b859eea6227cf71899) (your suggestion) - 1.7 seconds

So I think that would be an improvement, but switching to `first()` would still be more efficient by a large enough margin to make it worthwhile.