@drbyte 2020-11-27
Thanks for the updated benchmarks. That's helpful.

> but switching to `first()` would still be more efficient

Maybe I've been staring at code too long, but how are these different?

- getPermissions()->first()
 ... where getPermissions is returning the same collection that getPermission() refers to

- getPermission()
  ... where it calls ->first() on the same collection