@davejamesmiller 2020-11-27
`getPermissions()` calls `filter()`, which scans through the entire array and returns a new collection (probably with 0 or 1 elements), then calls `first()` on that. So PHP still has to allocate a new array and a new object for every call.

`getPermission()` calls `first()` on the original collection, which stops as soon as a match is found and never creates a new collection or array.