@davejamesmiller 2021-03-29
@drbyte Would you consider revisiting this? I still think it's needed. For now I have to override the class in every project. Thanks.

PS The bot says "feel free to reopen it" but we don't have permission to do that.