@drbyte 2021-03-29
@davejamesmiller It's still on my radar. I've been meaning to revisit it in conjunction with other PRs and issues talking about cache and performance. My worry is that the various approaches don't address each others' problems and might even work against each other.  Given that I haven't run into big issues with these myself I haven't staged an environment that addresses them all.