@drbyte 2021-08-17
@davejamesmiller much of your suggestion was incorporated into #1799 which has been merged and released in `4.3.0`.
Thanks for your contribution!